Create tab folder and create main tab file
