+++ fsleyes_plugin_shimming_toolbox/st_plugin.py
@@ -21,10 +21,9 @@ import fsleyes.views.canvaspanel as canvaspanel
 import logging
 import os
 from pathlib import Path
-import webbrowser
 import wx
 
-from fsleyes_plugin_shimming_toolbox import __dir_st_plugin__
+from fsleyes_plugin_shimming_toolbox.tabs.tab import Tab
 from fsleyes_plugin_shimming_toolbox.components.dropdown import DropdownComponent
 from fsleyes_plugin_shimming_toolbox.components.run import RunComponent
 from fsleyes_plugin_shimming_toolbox.components.input import InputComponent
@@ -43,13 +42,6 @@ CURR_DIR = os.getcwd()
 ST_DIR = f"{HOME_DIR}/shimming-toolbox"
 
 VERSION = "0.1.1"
-
-# Load icon resources
-rtd_logo = wx.Bitmap(os.path.join(__dir_st_plugin__, 'img', 'RTD.png'), wx.BITMAP_TYPE_PNG)
-# Load ShimmingToolbox logo saved as a png image, rescale it, and return it as a wx.Bitmap image.
-st_logo = wx.Image(os.path.join(__dir_st_plugin__, 'img', 'shimming_toolbox_logo.png'), wx.BITMAP_TYPE_PNG)
-st_logo.Rescale(int(st_logo.GetWidth() * 0.2), int(st_logo.GetHeight() * 0.2), wx.IMAGE_QUALITY_HIGH)
-st_logo = st_logo.ConvertToBitmap()
 
 
 # We need to create a ctrlpanel.ControlPanel instance so that it can be recognized as a plugin by FSLeyes
@@ -104,38 +96,6 @@ class NotebookTerminal(wx.Notebook):
         self.terminal_component = Terminal(parent)
 
 
-class Tab(wx.ScrolledWindow):
-    def __init__(self, parent, title, description):
-        super().__init__(parent)
-        self.title = title
-        self.sizer_info = InfoSection(self, description).sizer
-        self.terminal_component = parent.terminal_component
-        self.SetScrollbars(1, 4, 1, 1)
-
-    def create_sizer(self):
-        """Create the parent sizer for the tab.
-
-        Tab is divided into 2 main sizers:
-            sizer_info | sizer_run
-        """
-        sizer = wx.BoxSizer(wx.HORIZONTAL)
-        sizer.Add(self.sizer_info, 0)
-        sizer.AddSpacer(20)
-        sizer.Add(self.sizer_run, 2)
-        return sizer
-
-    def create_sizer_run(self):
-        """Create the run sizer containing tab-specific functionality."""
-        sizer = wx.BoxSizer(wx.VERTICAL)
-        sizer.SetMinSize(400, 300)
-        sizer.AddSpacer(10)
-        return sizer
-
-    def create_empty_component(self):
-        component = InputComponent(panel=self, input_text_box_metadata=[])
-        return component
-
-
 class Terminal:
     """Create the terminal where messages are logged to the user."""
     def __init__(self, panel):
@@ -153,38 +113,6 @@ class Terminal:
             self.terminal.AppendText(f"{msg}\n")
         else:
             self.terminal.AppendText(f"{level}: {msg}\n")
-
-
-class InfoSection:
-    def __init__(self, panel, description):
-        self.panel = panel
-        self.description = description
-        self.sizer = self.create_sizer()
-
-    def create_sizer(self):
-        """Create the left sizer containing generic Shimming Toolbox information."""
-        sizer = wx.BoxSizer(wx.VERTICAL)
-        logo = wx.StaticBitmap(parent=self.panel, id=-1, bitmap=st_logo, pos=wx.DefaultPosition)
-        width = logo.Size[0]
-        sizer.Add(logo, flag=wx.SHAPED, proportion=1)
-        sizer.AddSpacer(10)
-
-        # Create a "Documentation" button that redirects towards https://shimming-toolbox.org/en/latest/
-        button_documentation = wx.Button(self.panel, label="Documentation")
-        button_documentation.Bind(wx.EVT_BUTTON, self.open_documentation_url)
-        button_documentation.SetBitmap(rtd_logo)
-        sizer.Add(button_documentation, flag=wx.EXPAND)
-        sizer.AddSpacer(10)
-
-        # Add a short description of what the current tab does
-        description_text = wx.StaticText(self.panel, id=-1, label=self.description)
-        description_text.Wrap(width)
-        sizer.Add(description_text)
-        return sizer
-
-    def open_documentation_url(self, event):
-        """Redirect ``button_documentation`` to the ``shimming-toolbox`` page."""
-        webbrowser.open('https://shimming-toolbox.org/en/latest/')
 
 
 class B0ShimTab(Tab):
